% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instance_countrycode.R
\name{get_instance_countrycode}
\alias{get_instance_countrycode}
\title{Geocode a URL}
\usage{
get_instance_countrycode(instance)
}
\arguments{
\item{instance}{character URL. Not vectorized!}
}
\value{
country code, character
}
\description{
Should work with any URL, not just Mastodon instances.
}
