% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instance_stats.R
\name{get_instance_stats}
\alias{get_instance_stats}
\title{Get Information on a Mastodon Server}
\usage{
get_instance_stats(instance, verbose = TRUE)
}
\arguments{
\item{instance}{URL of the Mastodon Server. Not vectorized!}

\item{verbose}{boolean (default TRUE) print diagnostic messages}
}
\value{
data.frame with Server information
}
\description{
Get Information on a Mastodon Server
}
