% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instance_users.R
\name{get_instance_users}
\alias{get_instance_users}
\title{Loops the instance's user directory}
\usage{
get_instance_users(instance, from = NULL, limit = 40)
}
\arguments{
\item{instance}{character, URL of the Mastodon server}

\item{from}{integer (default NULL) offset. If NULL, the whole directory is parsed}

\item{limit}{integer (default 40). How many profiles to fetch in one request.}
}
\value{
data.frame with user information
}
\description{
Loops the instance's user directory
}
