% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instances.R
\name{get_instances}
\alias{get_instances}
\title{Return a manually curated list of Mastodon instances}
\usage{
get_instances()
}
\value{
character of URLs
}
\description{
Return a manually curated list of Mastodon instances
}
