% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_html.R
\name{strip_html}
\alias{strip_html}
\title{Remove HTML tags from a snippet}
\usage{
strip_html(x)
}
\arguments{
\item{x}{character}
}
\value{
character without HTML tags
}
\description{
See also https://stackoverflow.com/a/34344957/216064
}
